+++ typings/mixpanel.d.ts
@@ -24,7 +24,6 @@ declare namespace Mixpanel {
     append(unionObject: any, onSuccess: () => void, onFail: (errors: string) => void): void;
     deleteUser(onSuccess: () => void, onFail: (errors: string) => void): void;
     increment(peopleProperties: any, onSuccess: () => void, onFail: (errors: string) => void): void;
-    initPushHandling(senderId: string, onSuccess: () => void, onFail: (errors: string) => void): void;
     set(peopleProperties: any, onSuccess: () => void, onFail: (errors: string) => void): void;
     setOnce(peopleProperties: any, onSuccess: () => void, onFail: (errors: string) => void): void;
     setPushId(pushId: string, onSuccess: () => void, onFail: (errors: string) => void): void;