+++ www/mixpanel.js
@@ -178,14 +178,6 @@ mixpanel.people.increment = function(peopleProperties, onSuccess, onFail) {
   exec(onSuccess, onFail, 'Mixpanel', 'people_increment', [peopleProperties]);
 };
 
-mixpanel.people.initPushHandling = function(senderId, onSuccess, onFail) {
-  if (!senderId || typeof senderId !== 'string') {
-    return onFail(errors.invalid('senderId', senderId));
-  }
-
-  exec(onSuccess, onFail, 'Mixpanel', 'people_initPushHandling', [senderId]);
-}
-
 mixpanel.people.set = function(peopleProperties, onSuccess, onFail) {
   if (!peopleProperties || (typeof peopleProperties === 'object' && Object.keys(peopleProperties).length === 0)) {
     return onFail(errors.invalid('properties', peopleProperties));
@@ -204,7 +196,7 @@ mixpanel.people.setOnce = function(peopleProperties, onSuccess, onFail) {
 
 /**
  * @param pushId is the token/id you get back when registering the device with the notification service
- *        for android - this is the GCM token
+ *        for android - this is the FCM token
  *        for ios - this is the APN token
  */
 mixpanel.people.setPushId = function(pushId, onSuccess, onFail) {